Move SendAlerts inside util package
